Adds instructions for enabling cloud sync
Cloud sync is out of scope for the project, however may be acceptable
depending on the threat model.

This provides instructions to run the cloud service for folks that
might want that alongside local file encryption.

Adds instructions for enabling cloud sync

Cloud sync is out of scope for the project, however may be acceptable
depending on the threat model.

This provides instructions to run the cloud service for folks that
might want that alongside local file encryption.